split test_reference_counting into two files
